move statistics stuff into its own file. add online statistics computation
