+++ mondey_backend/src/mondey_backend/routers/utils.py
@@ -4,7 +4,6 @@ import datetime
 import logging
 import pathlib
 from collections.abc import Iterable
-from collections.abc import Sequence
 from typing import TypeVar
 
 import numpy as np
@@ -23,14 +22,10 @@ from ..models.children import Child
 from ..models.milestones import AgeInterval
 from ..models.milestones import Milestone
 from ..models.milestones import MilestoneAdmin
-from ..models.milestones import MilestoneAgeScore
-from ..models.milestones import MilestoneAgeScoreCollection
 from ..models.milestones import MilestoneAnswer
 from ..models.milestones import MilestoneAnswerSession
 from ..models.milestones import MilestoneGroup
 from ..models.milestones import MilestoneGroupAdmin
-from ..models.milestones import MilestoneGroupAgeScore
-from ..models.milestones import MilestoneGroupAgeScoreCollection
 from ..models.milestones import MilestoneGroupText
 from ..models.milestones import MilestoneText
 from ..models.questions import ChildQuestion
@@ -214,189 +209,6 @@ def _get_answer_session_child_ages_in_months(session: SessionDep) -> dict[int, i
 def _get_expected_age_from_scores(scores: np.ndarray) -> int:
     # placeholder algorithm: returns first age with avg score > 3
     return np.argmax(scores >= 3.0)
-
-
-def _get_statistics_by_age(
-    answers: Sequence[MilestoneAnswer], child_ages: dict[int, int]
-) -> tuple[np.ndarray, np.ndarray]:
-    """
-    _summary_
-
-    Parameters
-    ----------
-    answers : Sequence[MilestoneAnswer]
-        _description_
-    child_ages : dict[int, int]
-        _description_
-
-    Returns
-    -------
-    tuple[np.ndarray, np.ndarray]
-        _description_
-    """
-    max_age_months = 72
-    avg_scores = np.zeros(max_age_months + 1)
-    stddev_scores = np.zeros(max_age_months + 1)
-    counts = np.zeros_like(avg_scores)
-    if child_ages == {}:
-        return avg_scores, stddev_scores
-
-    # compute average
-    for answer in answers:
-        age = child_ages[answer.answer_session_id]  # type: ignore
-        # convert 0-3 answer index to 1-4 score
-        avg_scores[age] += answer.answer + 1
-        counts[age] += 1
-
-    with np.errstate(invalid="ignore"):
-        avg_scores /= counts
-
-    # compute standard deviation
-    for answer in answers:
-        age = child_ages[answer.answer_session_id]  # type: ignore
-        stddev_scores[age] += ((answer.answer + 1) - avg_scores[age]) ** 2
-
-    with np.errstate(invalid="ignore"):
-        stddev_scores = np.sqrt(stddev_scores / np.max(counts - 1, 0))
-
-    # replace NaNs (due to zero counts) with zeros
-    avg = np.nan_to_num(avg_scores)
-    stddev = np.nan_to_num(stddev_scores)
-
-    return avg, stddev
-
-
-def calculate_milestone_statistics_by_age(
-    session: SessionDep,
-    milestone_id: int,
-    answers: Sequence[MilestoneAnswer] | None = None,
-) -> MilestoneAgeScoreCollection:
-    """
-    _summary_
-
-    Parameters
-    ----------
-    session : SessionDep
-        _description_
-    milestone_id : int
-        _description_
-    answers : Sequence[MilestoneAnswer] | None, optional
-        _description_, by default None
-
-    Returns
-    -------
-    MilestoneAgeScoreCollection
-        _description_
-    """
-    child_ages = _get_answer_session_child_ages_in_months(session)
-    # FIXME: change this to an online algorithm that starts with the last known statistcs
-    # and then adds the new answers
-    if answers is None:
-        answers = session.exec(
-            select(MilestoneAnswer).where(
-                col(MilestoneAnswer.milestone_id) == milestone_id
-            )
-        ).all()
-
-    avg, stddev = _get_statistics_by_age(answers, child_ages)
-    expected_age = _get_expected_age_from_scores(avg)
-    return MilestoneAgeScoreCollection(
-        id=None,
-        milestone_id=milestone_id,
-        expected_age=expected_age,
-        created_at=datetime.datetime.now(),
-        scores=[
-            MilestoneAgeScore(
-                id=None,
-                collection_id=None,
-                avg_score=avg[age],
-                stddev_score=stddev[age],
-                age_months=age,
-                expected_score=4
-                if age >= expected_age
-                else 1,  # TODO: placeholder algorithm? how does the model behind this look like really?
-            )
-            for age in range(0, len(avg))
-        ],
-    )
-
-
-def calculate_milestonegroup_statistics_by_age(
-    session: SessionDep,
-    milestonegroup_id,
-    answers: Sequence[MilestoneAnswer] | None = None,
-) -> MilestoneGroupAgeScoreCollection:
-    """
-    _summary_
-
-    Parameters
-    ----------
-    session : SessionDep
-        _description_
-    milestonegroup_id : _type_
-        _description_
-    answers : Sequence[MilestoneAnswer] | None, optional
-        _description_, by default None
-
-    Returns
-    -------
-    MilestoneGroupAgeScoreCollection
-        _description_
-    """
-    # FIXME: change this to an online algorithm that starts with the last known statistcs
-    # and then adds the new answers
-    child_ages = _get_answer_session_child_ages_in_months(session)
-
-    if answers is None:
-        answers = session.exec(
-            select(MilestoneAnswer).where(
-                col(MilestoneAnswer.milestone_group_id) == milestonegroup_id
-            )
-        ).all()
-
-    avg, stddev = _get_statistics_by_age(answers, child_ages)
-    return MilestoneGroupAgeScoreCollection(
-        id=None,
-        milestonegroup_id=milestonegroup_id,
-        scores=[
-            MilestoneGroupAgeScore(
-                id=None,
-                collection_id=None,
-                age_months=age,
-                avg_score=avg[age],
-                stddev_score=stddev[age],
-                milestonegroup_id=milestonegroup_id,
-            )
-            for age in range(0, len(avg))
-        ],
-        created_at=datetime.datetime.now(),
-    )
-
-
-def recompute_milestonegroup_statistics(session: SessionDep):
-    # fetch all milestonegroup statsitcs and check how old they are. Then
-    # recompute the ones that are older than timedelta and put back into database
-    # do the same for milestone statistics
-
-    milestonegroups = session.exec(select(MilestoneGroup.id)).all()
-    for milestonegroup in milestonegroups:
-        statistics = calculate_milestonegroup_statistics_by_age(session, milestonegroup)
-        for score in statistics.scores:
-            add(session, score)
-        add(session, statistics)
-
-
-def recompute_milestone_statistics(session: SessionDep):
-    # fetch all milestonegroup statsitcs and check how old they are. Then
-    # recompute the ones that are older than timedelta and put back into database
-    # do the same for milestone statistics
-
-    milestones = session.exec(select(Milestone.id)).all()
-    for milestone in milestones:
-        statistics = calculate_milestone_statistics_by_age(session, milestone)  # type: ignore
-        for score in statistics.scores:
-            add(session, score)
-        add(session, statistics)
 
 
 def child_image_path(child_id: int | None) -> pathlib.Path:
